requirements.txt
Flask==2.3.3
qrcode==7.4.2
Pillow==10.0.1
mysql-connector-python==8.0.33